feat(bash/lib): improve VS Code Remote Tunnels/SSH detection
Currently, it works by getting the VSCODE_GIT_ASKPASS_NODE variable first
as an argument for dirname for the parent directory, and the appending that
output to point to the remote CLI binary. I might improve this further by
not relying on the default /home/<username> path and utilize the HOME
variable.

Signed-off-by: Andrei Jiroh Halili <ajhalili2006@andreijiroh.dev>